Backport 1.12.2
fix


fix2


fix


fix2


fix config, fix optifine


fix missing lib


fix


make simpler


fix mixin json


fix


fix2


fix3


fix


fix2


hopefully fix spaces in file name


is this file even used on 1.12?


fix playSound on old mc


fix


fix


fix player nametags


fix getEntities on -=1.12.2

+++ common/src/main/java/xyz/wagyourtail/jsmacros/client/access/IInventory.java
@@ -1,8 +1,7 @@
 package xyz.wagyourtail.jsmacros.client.access;
 
-
-import net.minecraft.container.Slot;
+import net.minecraft.inventory.slot.Slot;
 
 public interface IInventory {
-    Slot jsmacros_getSlotUnder(double x, double y);
+    Slot jsmacros_getSlotUnder(int x, int y);
 }
